+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/db/entity/Account.kt
@@ -12,10 +12,5 @@ data class Account(
     val account: Int? = 0,
 
     @ColumnInfo(name = "ufvk")
-    val unifiedFullViewingKey: String? = "",
-
-    val address: String? = "",
-
-    @ColumnInfo(name = "transparent_address")
-    val transparentAddress: String? = ""
+    val unifiedFullViewingKey: String? = ""
 )